moved basic info into form for now
